Move new functions to own file
